+++ FieldOpt/Simulation/simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/schedule_section.h
@@ -43,7 +43,7 @@ class Schedule : public ECLDriverPart
 {
  public:
   Schedule(QList<Model::Wells::Well *> *wells,
-           QList<int> control_times);
+           QList<double> control_times);
   QString GetPartString();
 
  private:
